\name{qdgl-deprecated}
\alias{qdgl-deprecated}
\alias{qdgl}
\title{Deprecated function for density quantile function of gld. See qdgl instead}
\usage{
qdgl(p, lambda1, lambda2 = NULL, lambda3 = NULL, lambda4 = NULL, 
  param = "fkml", lambda5 = NULL)
}
\arguments{
  \item{p}{See qdgl help instead.}
  \item{lambda1}{See qdgl help instead.}
  \item{lambda2}{See qdgl help instead.}
  \item{lambda3}{See qdgl help instead.}
  \item{lambda4}{See qdgl help instead.}
  \item{param}{See qdgl help instead.}
  \item{lambda5}{See qdgl help instead.}
}
\description{See qdgl help instead.}
\value{See qdgl help instead.}
